Added persistence test to event service.
* Added PersistenceTests test class with
crud tests and failure tests.
* Added configuration class for event entity
 validation annotation detection.
* Modified event entity data model, added
EventId attribute, to facilitate testing.

+++ events-service/src/main/java/com/calendar/service/event/persistence/EventRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 
 public interface EventRepository extends CrudRepository<Event, Long> {
-	Optional<Event> findByEventId(int eventId);
+	Optional<Event> findByEventId(Long eventId);
 }
